+++ mobile/touches_empty.go
@@ -21,5 +21,5 @@ package mobile
 func updateTouchesOnAndroid(action int, id int, x, y int) {
 }
 
-func updateTouchesOnIOSImpl(phase int, ptr int, x, y int) {
+func updateTouchesOnIOSImpl(phase int, ptr int64, x, y int) {
 }
